+++ cli/csstools-cli/src/plugins/postcss-light-dark-function.ts
@@ -1,0 +1,18 @@
+import plugin from '@csstools/postcss-light-dark-function';
+import { cli, helpTextLogger } from '@csstools/base-cli';
+
+export default function postcssLightDarkFunction() {
+	cli(
+		plugin,
+		['preserve'],
+		helpTextLogger(
+			'@csstools/cli postcss-light-dark-function',
+			'PostCSS Light Dark function',
+			'Use the light-dark() color function in CSS',
+			{
+				preserve: true,
+			},
+		),
+		false,
+	);
+}