+++ e2e-package-managers/yarn/test.sh
@@ -2,5 +2,6 @@
 
 set -e;
 
+touch yarn.lock;
 yarn install;
 node ./test.mjs;