+++ packages/css-tokenizer/src/consume/numeric-token.ts
@@ -20,7 +20,7 @@ export function consumeNumericToken(ctx: Context, reader: CodePointReader): Toke
 	}
 
 	const numberType = consumeNumber(ctx, reader);
-	const numberValue = parseFloat(reader.source.slice(reader.representationStart, reader.representationEnd + 1)) || 0;
+	const numberValue = parseFloat(reader.source.slice(reader.representationStart, reader.representationEnd + 1));
 
 	if (checkIfThreeCodePointsWouldStartAnIdentSequence(ctx, reader)) {
 		const unit = consumeIdentSequence(ctx, reader);