+++ packages/utilities/test/_require.cjs
@@ -1,0 +1,4 @@
+const assert = require('assert');
+const { hasFallback } = require('@csstools/utilities');
+
+assert.ok(typeof hasFallback === 'function');