+++ plugins/postcss-nesting/src/editions/2024-02/index.ts
@@ -1,0 +1,34 @@
+import type { PluginCreator } from 'postcss';
+import ampersandToScope from './lib/ampersand-to-scope.js';
+import walk from './lib/walk.js';
+
+/** postcss-nesting plugin options */
+export type pluginOptions = {
+	/** @deprecated This option was removed. You must migrate your CSS to the latest speciation to continue using this plugin. */
+	noIsPseudoSelector?: boolean,
+};
+
+const creator: PluginCreator<pluginOptions> = () => {
+	return {
+		postcssPlugin: 'postcss-nesting',
+		Rule(rule, { result }) {
+			walk(rule, result);
+
+			if (rule.selector.includes('&')) {
+				ampersandToScope(rule, result);
+			}
+		},
+		AtRule: {
+			nest(rule) {
+				throw rule.error(
+					'`@nest` was removed from the CSS Nesting specification and will be removed from PostCSS Nesting in the next major version.\n' +
+					`Change \`@nest ${rule.params} {}\` to \`${rule.params} {}\` to migrate to the latest standard.`,
+				);
+			},
+		},
+	};
+};
+
+creator.postcss = true;
+
+export default creator;