+++ plugins/postcss-nesting/src/index.ts
@@ -1,36 +1,37 @@
 import type { PluginCreator } from 'postcss';
-import ampersandToScope from './lib/ampersand-to-scope.js';
-import walk from './lib/walk.js';
+import type { pluginOptions as pluginOptions2021 } from './editions/2021';
+import type { pluginOptions as pluginOptions2024_02 } from './editions/2024-02';
+import plugin2021 from './editions/2021';
+import plugin2024_02 from './editions/2024-02';
+
+export type { pluginOptions as pluginOptions2021 } from './editions/2021';
+export type { pluginOptions as pluginOptions2024_02 } from './editions/2024-02';
+
 
 /** postcss-nesting plugin options */
 export type pluginOptions = {
-	/** Avoid the `:is()` pseudo class as much as possible. default: false */
-	noIsPseudoSelector?: boolean,
-	/** Silence the `@nest` warning. */
-	silenceAtNestWarning?: boolean,
-};
+	/** The implementation edition for CSS Nesting, default to '2021' */
+	edition?: '2021' | '2024-02',
+} & pluginOptions2021 & pluginOptions2024_02;
 
 const creator: PluginCreator<pluginOptions> = (opts?: pluginOptions) => {
 	const options = Object.assign(
 		// Default options
 		{
-			noIsPseudoSelector: false,
-			silenceAtNestWarning: false,
+			edition: '2021',
 		},
 		// Provided options
 		opts,
 	);
 
-	return {
-		postcssPlugin: 'postcss-nesting',
-		Rule(rule, { result }) {
-			walk(rule, result, options);
-
-			if (rule.selector.includes('&')) {
-				ampersandToScope(rule, result);
-			}
-		},
-	};
+	switch (options.edition) {
+		case '2021':
+			return plugin2021(opts);
+		case '2024-02':
+			return plugin2024_02(opts);
+		default:
+			throw new Error(`Invalid edition: ${options.edition}`);
+	}
 };
 
 creator.postcss = true;