+++ packages/cascade-layer-name-parser/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changes to Cascade Layer Name Parser
 
+### 1.0.8
+
+_February 19, 2024_
+
+- Updated [`@csstools/css-parser-algorithms`](https://github.com/csstools/postcss-plugins/tree/main/packages/css-parser-algorithms) to [`2.6.0`](https://github.com/csstools/postcss-plugins/tree/main/packages/css-parser-algorithms/CHANGELOG.md#260) (minor)
+
 ### 1.0.7
 
 _December 31, 2023_