+++ packages/media-query-list-parser/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changes to Media Query List Parser
 
+### 2.1.8
+
+_February 19, 2024_
+
+- Updated [`@csstools/css-parser-algorithms`](https://github.com/csstools/postcss-plugins/tree/main/packages/css-parser-algorithms) to [`2.6.0`](https://github.com/csstools/postcss-plugins/tree/main/packages/css-parser-algorithms/CHANGELOG.md#260) (minor)
+
 ### 2.1.7
 
 _December 31, 2023_