+++ plugins/postcss-color-rebeccapurple/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changes to PostCSS RebeccaPurple
 
+### 9.0.3
+
+_February 19, 2024_
+
+- Reuse code with `@csstools/utilities`
+
 ### 9.0.2
 
 _December 15, 2023_