+++ plugins/postcss-custom-properties/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changes to PostCSS Custom Properties
 
+### 13.3.5
+
+_February 19, 2024_
+
+- Reuse code with `@csstools/utilities`
+- Updated [`@csstools/css-parser-algorithms`](https://github.com/csstools/postcss-plugins/tree/main/packages/css-parser-algorithms) to [`2.6.0`](https://github.com/csstools/postcss-plugins/tree/main/packages/css-parser-algorithms/CHANGELOG.md#260) (minor)
+- Updated [`@csstools/cascade-layer-name-parser`](https://github.com/csstools/postcss-plugins/tree/main/packages/cascade-layer-name-parser) to [`1.0.8`](https://github.com/csstools/postcss-plugins/tree/main/packages/cascade-layer-name-parser/CHANGELOG.md#108) (patch)
+
 ### 13.3.4
 
 _December 31, 2023_