+++ plugins/postcss-double-position-gradients/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changes to PostCSS Double Position Gradients
 
+### 5.0.4
+
+_February 19, 2024_
+
+- Reuse code with `@csstools/utilities`
+- Updated [`@csstools/postcss-progressive-custom-properties`](https://github.com/csstools/postcss-plugins/tree/main/plugins/postcss-progressive-custom-properties) to [`3.1.0`](https://github.com/csstools/postcss-plugins/tree/main/plugins/postcss-progressive-custom-properties/CHANGELOG.md#310) (minor)
+
 ### 5.0.3
 
 _December 15, 2023_