+++ plugins/postcss-gradients-interpolation-method/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changes to PostCSS Gradients Interpolation Method
 
+### 4.0.11
+
+_February 26, 2024_ 
+
+- Improve the regex for gradient function detection
+
+### 4.0.10
+
+_February 19, 2024_
+
+- Reuse code with `@csstools/utilities`
+- Updated [`@csstools/css-parser-algorithms`](https://github.com/csstools/postcss-plugins/tree/main/packages/css-parser-algorithms) to [`2.6.0`](https://github.com/csstools/postcss-plugins/tree/main/packages/css-parser-algorithms/CHANGELOG.md#260) (minor)
+- Updated [`@csstools/postcss-progressive-custom-properties`](https://github.com/csstools/postcss-plugins/tree/main/plugins/postcss-progressive-custom-properties) to [`3.1.0`](https://github.com/csstools/postcss-plugins/tree/main/plugins/postcss-progressive-custom-properties/CHANGELOG.md#310) (minor)
+- Updated [`@csstools/css-color-parser`](https://github.com/csstools/postcss-plugins/tree/main/packages/css-color-parser) to [`1.5.2`](https://github.com/csstools/postcss-plugins/tree/main/packages/css-color-parser/CHANGELOG.md#152) (patch)
+
 ### 4.0.9
 
 _December 31, 2023_