+++ plugins/postcss-image-set-function/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changes to PostCSS image-set() Function
 
+### 6.0.3
+
+_February 19, 2024_
+
+- Reuse code with `@csstools/utilities`
+
 ### 6.0.2
 
 _December 15, 2023_