+++ plugins/postcss-is-pseudo-class/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changes to PostCSS Is Pseudo Class
 
+### 4.0.5
+
+_February 19, 2024_
+
+- Updated [`@csstools/selector-specificity`](https://github.com/csstools/postcss-plugins/tree/main/packages/selector-specificity) to [`3.0.2`](https://github.com/csstools/postcss-plugins/tree/main/packages/selector-specificity/CHANGELOG.md#302) (patch)
+
 ### 4.0.4
 
 _December 15, 2023_