+++ plugins/postcss-media-queries-aspect-ratio-number-values/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changes to PostCSS Media Queries Aspect-Ratio Number Values
 
+### 2.0.6
+
+_February 19, 2024_
+
+- Updated [`@csstools/css-parser-algorithms`](https://github.com/csstools/postcss-plugins/tree/main/packages/css-parser-algorithms) to [`2.6.0`](https://github.com/csstools/postcss-plugins/tree/main/packages/css-parser-algorithms/CHANGELOG.md#260) (minor)
+- Updated [`@csstools/media-query-list-parser`](https://github.com/csstools/postcss-plugins/tree/main/packages/media-query-list-parser) to [`2.1.8`](https://github.com/csstools/postcss-plugins/tree/main/packages/media-query-list-parser/CHANGELOG.md#218) (patch)
+
 ### 2.0.5
 
 _December 31, 2023_